make bootstrap-crostini a little smarter and idempotent

+++ bootstrap-crostini.sh
@@ -2,31 +2,43 @@
 source ./bootstrap-lib.sh
 
 # tailscale
-curl -fsSL https://pkgs.tailscale.com/stable/debian/bookworm.noarmor.gpg | sudo tee /usr/share/keyrings/tailscale-archive-keyring.gpg >/dev/null
-curl -fsSL https://pkgs.tailscale.com/stable/debian/bookworm.tailscale-keyring.list | sudo tee /etc/apt/sources.list.d/tailscale.list
-sudo apt update
-sudo apt install -y tailscale
+dpkg -l tailscale > /dev/null 2>&1
+if [[ $? -ne 0 ]]; then
+    curl -fsSL https://pkgs.tailscale.com/stable/debian/bookworm.noarmor.gpg | sudo tee /usr/share/keyrings/tailscale-archive-keyring.gpg >/dev/null
+    curl -fsSL https://pkgs.tailscale.com/stable/debian/bookworm.tailscale-keyring.list | sudo tee /etc/apt/sources.list.d/tailscale.list
+    sudo apt update
+    sudo apt install -y tailscale
+fi
+
 
 # docker - because podman doesn't work in crostini/termina
-sudo apt install -y ca-certificates curl gnupg
-sudo install -m 0755 -d /etc/apt/keyrings
-curl -fsSL https://download.docker.com/linux/debian/gpg | sudo gpg --dearmor -o /etc/apt/keyrings/docker.gpg --yes
-sudo chmod a+r /etc/apt/keyrings/docker.gpg
-echo \
-  "deb [arch="$(dpkg --print-architecture)" signed-by=/etc/apt/keyrings/docker.gpg] https://download.docker.com/linux/debian \
-  "$(. /etc/os-release && echo "$VERSION_CODENAME")" stable" | \
-  sudo tee /etc/apt/sources.list.d/docker.list > /dev/null
-sudo apt update
-sudo apt install -y docker-ce docker-ce-cli containerd.io docker-buildx-plugin docker-compose-plugin
-sudo groupadd docker
-sudo usermod -aG docker $USER
+dpkg -l docker-ce > /dev/null 2>&1
+if [[ $? -ne 0 ]]; then
+    sudo apt install -y ca-certificates curl gnupg
+    sudo install -m 0755 -d /etc/apt/keyrings
+    curl -fsSL https://download.docker.com/linux/debian/gpg | sudo gpg --dearmor -o /etc/apt/keyrings/docker.gpg --yes
+    sudo chmod a+r /etc/apt/keyrings/docker.gpg
+    echo \
+      "deb [arch="$(dpkg --print-architecture)" signed-by=/etc/apt/keyrings/docker.gpg] https://download.docker.com/linux/debian \
+      "$(. /etc/os-release && echo "$VERSION_CODENAME")" stable" | \
+      sudo tee /etc/apt/sources.list.d/docker.list > /dev/null
+    sudo apt update
+    sudo apt install -y docker-ce docker-ce-cli containerd.io docker-buildx-plugin docker-compose-plugin
+    sudo groupadd docker
+    sudo usermod -aG docker $USER
+fi
 
 # rootless distrobox
-curl -s https://raw.githubusercontent.com/89luca89/distrobox/main/install | sh -s -- --prefix ~/.local
+if [[ ! -f ${HOME}/.local/bin/distrobox ]]; then
+    curl -s https://raw.githubusercontent.com/89luca89/distrobox/main/install | sh -s -- --prefix ~/.local
+fi
 
 # nodejs LTS
-curl -fsSL https://deb.nodesource.com/setup_18.x | sudo -E bash - &&\
-sudo apt install -y nodejs
+dpkg -l nodejs | grep 18\. > /dev/null 2>&1
+if [[ $? -ne 0 ]]; then
+    curl -fsSL https://deb.nodesource.com/setup_18.x | sudo -E bash - &&\
+    sudo apt install -y nodejs
+fi
 
 # ssh-agent systemd user unit
 mkdir_if_needed ~/.config/systemd/user
@@ -50,15 +62,21 @@ systemctl --user enable ssh-agent
 sudo apt install -y vim python3 python3-pip python3-venv python3-q \
     git tmux htop strace pipx virtualenvwrapper
 
-# download golang
+# golang
 golang_version="1.20.7"
-sudo apt remove -y golang 
-if [[ -d /usr/local/go ]]; then
+dpkg -l golang > /dev/null 2>&1
+if [[ $? -eq 0 ]]; then
+    sudo apt remove -y golang 
+fi
+go version | grep "$golang_version" > /dev/null 2>&1
+if [[ $? -ne 0 ]]; then
     sudo rm -fr /usr/local/go
 fi
-sudo curl -o "/usr/local/go-${golang_version}.tar.gz" "https://dl.google.com/go/go${golang_version}.linux-$(dpkg --print-architecture).tar.gz"
-sudo tar -zxvf /usr/local/go-${golang_version}.tar.gz --directory=/usr/local/
-sudo rm /usr/local/go-${golang_version}.tar.gz
+if [[ ! -d /usr/local/go ]]; then
+    sudo curl -o "/usr/local/go-${golang_version}.tar.gz" "https://dl.google.com/go/go${golang_version}.linux-$(dpkg --print-architecture).tar.gz"
+    sudo tar -zxvf /usr/local/go-${golang_version}.tar.gz --directory=/usr/local/
+    sudo rm /usr/local/go-${golang_version}.tar.gz
+fi
 
 for pypkg in ptpython tox httpie flake8 pep8 pyflakes pylint black pipenv poetry tmuxp bpytop python-lsp-server tldr
 do
